Moved suite_id interface to network/
