+++ lib/cylc/network/suite_id.py
